list: new file
move here all the code for listing containers.
